move apply functionality from BaseHandler to SimpleHandler
The only other child class WriteHandler uses a different implementation
for apply, so the generic variant is pointless.
